orders/src/events/listeners/__test__/ticket-created-listener.test.ts
import { Message } from "node-nats-streaming";
import { TicketCreatedEvent } from "@bala-tickets/common";
import { natsWrapper } from "../../../nats-wrapper";
import { TicketCreatedListener } from "../ticket-created-listener";
import mongoose from "mongoose";
import { Ticket } from "../../../models/ticket";

const setup = async () => {
	// create an instance of the listener
	const listener = new TicketCreatedListener(natsWrapper.client);

	// create a fake data event
	const data: TicketCreatedEvent["data"] = {
		version: 0,
		id: new mongoose.Types.ObjectId().toHexString(),
		title: "concert",
		price: 10,
		userId: new mongoose.Types.ObjectId().toHexString(),
	};

	// create a fake message object
	// @ts-ignore
	const message: Message = {
		ack: jest.fn(),
	};

	return { listener, data, message };
};

it("creates and saves a ticket", async () => {
    const {listener, data, message} = await setup();

	// call the onMessage function with the data object + message object
    await listener.onMessage(data, message);

	// write assertions to make sure a ticket was created!
    const ticket = await Ticket.findById(data.id);

    expect(ticket).toBeDefined();
    expect(ticket!.title).toEqual(data.title);
    expect(ticket!.price).toEqual(data.price);
});

it("ack the message", async () => {
    const {listener, data, message} = await setup();

	// call the onMessage function with the data object + message object
    await listener.onMessage(data, message);

	// write assertions to make sure ack function is called
    expect(message.ack).toBeCalled();
});
